move full game test into the tournament object
