+++ trials-engine/src/test/java/org/gardler/biglittlechallenge/trials_engine/TournamentTest.java
@@ -1,7 +1,13 @@
 package org.gardler.biglittlechallenge.trials_engine;
 
+import static org.junit.Assert.assertEquals;
+
 import org.gardler.biglittlechallenge.core.model.AbstractGame;
+import org.gardler.biglittlechallenge.core.model.GameStatus;
+import org.gardler.biglittlechallenge.trials.EngineControllerApp;
 import org.gardler.biglittlechallenge.trials.engine.Tournament;
+import org.gardler.biglittlechallenge.trials.engine.TournamentAPI;
+import org.gardler.biglittlechallenge.trials.model.Player;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -18,5 +24,22 @@ public class TournamentTest extends TestCase {
 		logger.info("Tournament toString is: " + result);
 		assertTrue("We don;t appear to have set the rounds in the tournament correctly", result.contains("consists of 4 events"));
 	}
-
+	
+	public void testFullGame() {
+		String uiClassName = "org.gardler.biglittlechallenge.trials.ai.DumbAIUI";
+		Player player = new Player("Test AI Player", uiClassName); 
+		
+		TournamentAPI api = new TournamentAPI();
+		api.abortGame(); // This is to ensure we are not using a previously created game
+		
+		GameStatus status = api.postJoinGame(player);
+		assertEquals("Game status is not 'WaitingForPlayers'", GameStatus.State.WaitingForPlayers, status.getState());
+		
+		player = new Player("Second Test AI Player", uiClassName);
+		status = api.postJoinGame(player);
+		assertEquals("Game status is not 'Playing' after second player joins", GameStatus.State.Playing, status.getState());
+		
+		
+		
+	}
 }
